requirements.txt
pip~=23.0.1
typing_extensions~=4.12.2
exceptiongroup~=1.2.2
sniffio~=1.3.1
starlette~=0.41.0
anyio~=4.6.2.post1
setuptools~=65.5.0
uvicorn~=0.32.0
aiopg
pydantic
aio-pika
aioredis
python-decouple
passlib
pyjwt
uvicorn